requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2
openai==0.28.1
httpx==0.27.0
